+++ docs/changelog.md
@@ -20,9 +20,10 @@
 - Documentation now reflects that the Hue component for some color types can be NaN
 - Static functions in the documentation now have a badge
 - Read-only properties in the documentation now have a badge
+- Updated the allowed interpolations for `Color.mix`
 - `Color.components` now allows you to obtain unclipped components
 - `Color.luminance` now corrects the [error](https://www.w3.org/WAI/GL/wiki/index.php?title=Relative_luminance&oldid=11187) from the equation provided in WCAG 2
-- `Color.mix` now accepts any color type except `Hex`
+- `Gradient.toColorSequence` was renamed to `Gradient.colorSequence`
 
 ## [0.1.0] - 2021-11-09
 ### Added